+++ src/program/Config.cpp
@@ -150,7 +150,6 @@ void Config::save(const std::string& gamepath) {
     }
     settings.endArray();
 
-    settings.setValue("save_screenpixels", sc.save_screenpixels);
     settings.setValue("savestate_settings", sc.savestate_settings);
 
     settings.endGroup();
@@ -269,7 +268,6 @@ void Config::load(const std::string& gamepath) {
     sc.video_framerate = settings.value("video_framerate", sc.video_framerate).toInt();
     sc.audio_codec = settings.value("audio_codec", sc.audio_codec).toInt();
     sc.audio_bitrate = settings.value("audio_bitrate", sc.audio_bitrate).toInt();
-    sc.save_screenpixels = settings.value("save_screenpixels", sc.save_screenpixels).toBool();
     sc.savestate_settings = settings.value("savestate_settings", sc.savestate_settings).toInt();
     sc.opengl_soft = settings.value("opengl_soft", sc.opengl_soft).toBool();
     sc.opengl_performance = settings.value("opengl_performance", sc.opengl_performance).toBool();