+++ src/program/ui/MainWindow.h
@@ -108,7 +108,6 @@ public:
     QActionGroup *timeSecGroup;
 
     QAction *busyloopAction;
-    QAction *saveScreenAction;
     QAction *preventSavefileAction;
     QAction *recycleThreadsAction;
 
@@ -254,7 +253,6 @@ private slots:
     void slotOsdEncode(bool checked);
 #endif
     void slotBusyLoop(bool checked);
-    void slotSaveScreen(bool checked);
     void slotPreventSavefile(bool checked);
     void slotMovieEnd();
     void slotPauseMovie();